Add troubleshooting step within cleaning up
Something that is not made clear in this guide is that the names of controllers in the kext need to be long enough for it to be injected. I discovered this by accident as I titled each controller with only 4 characters, except one, which I left I note to finish it up in the title and forgot to remove. Only that controller was mapped when I finished the map. I theorized it was length that was the issue and tested my theory. After loading the updated kext, the map was successfully applied.